[Dy2St] Move some transformer specific utilities into transformer dir
