+++ python/paddle/jit/dy2static/transformers/break_continue_transformer.py
@@ -13,11 +13,11 @@
 # limitations under the License.
 
 from paddle.base import unique_name
-from paddle.jit.dy2static.utils import BaseNodeVisitor, index_in_list
-from paddle.jit.dy2static.variable_trans_func import create_bool_node
+from paddle.jit.dy2static.utils import index_in_list
 from paddle.utils import gast
 
 from .base import BaseTransformer, ForNodeVisitor
+from .utils import BaseNodeVisitor, create_bool_node
 
 __all__ = []
