+++ python/paddle/jit/dy2static/utils.py
@@ -16,7 +16,6 @@ from __future__ import annotations
 
 import atexit
 import builtins
-import copy
 import functools
 import importlib.util
 import inspect
@@ -94,35 +93,6 @@ NO_SHAPE_VAR_TYPE = [
     core.VarDesc.VarType.FEED_MINIBATCH,
     core.VarDesc.VarType.FETCH_LIST,
 ]
-
-
-class BaseNodeVisitor(gast.NodeVisitor):
-    """
-    Implement customized NodeVisitor inherited from gast.NodeVisitor.
-    Ancestor nodes are traced to easily support more operations of currently
-    visited node.
-    """
-
-    def __init__(self):
-        self.ancestor_nodes = []
-
-    def visit(self, node):
-        """Visit a node."""
-        self.ancestor_nodes.append(node)
-
-        method = 'visit_' + node.__class__.__name__
-        visitor = getattr(self, method, self.generic_visit)
-        ret = visitor(node)
-        self.ancestor_nodes.pop()
-        return ret
-
-
-def get_parent_mapping(root):
-    to_parent: dict[gast.AST, gast.AST] = {}
-    for node in gast.walk(root):
-        for child in gast.iter_child_nodes(node):
-            to_parent[child] = node
-    return to_parent
 
 
 def data_layer_not_check(name, shape, dtype='float32', lod_level=0):
@@ -310,110 +280,6 @@ def is_paddle_func(func, ignore_white_list=True):
         return flag
     except Exception:
         return False
-
-
-def create_api_shape_node(tensor_shape_node):
-    assert isinstance(
-        tensor_shape_node, (gast.Name, gast.Attribute, gast.Subscript)
-    )
-
-    if isinstance(tensor_shape_node, gast.Name):
-        api_shape_node = gast.Call(
-            func=gast.parse('paddle.shape').body[0].value,
-            args=[tensor_shape_node],
-            keywords=[],
-        )
-        return api_shape_node
-
-    if isinstance(tensor_shape_node, gast.Attribute):
-        api_shape_node = gast.Call(
-            func=gast.parse('paddle.shape').body[0].value,
-            args=[tensor_shape_node.value],
-            keywords=[],
-        )
-        return api_shape_node
-
-    if isinstance(tensor_shape_node, gast.Subscript):
-        result_node = copy.deepcopy(tensor_shape_node)
-        result_node.value = create_api_shape_node(result_node.value)
-        return result_node
-
-
-def get_constant_variable_node(name, value, shape=[1], dtype='int64'):
-    return gast.parse(
-        f'{name} = paddle.full({str(shape)}, "{str(value)}", {dtype})'
-    )
-
-
-def get_attribute_full_name(node):
-    assert isinstance(
-        node, gast.Attribute
-    ), "Input non-Attribute node to get attribute full name"
-    return ast_to_source_code(node).strip()
-
-
-def generate_name_node(name_ids, ctx=gast.Load(), gen_tuple_if_single=False):
-    """
-    If name_ids is list or tuple or set with multiple strings, this function
-    generates gast.Tuple of gast.Name.
-    If the name_ids is single string or contains only 1 string, this function
-    returns gast.Name if gen_tuple_if_single==False else returns gast.Tuple
-    with only one gast.Name
-
-    This function is used at several gast.Return statements.
-    """
-    if isinstance(name_ids, str):
-        name_ids = [name_ids]
-    if not isinstance(name_ids, (list, tuple, set)):
-        raise TypeError(
-            'name_ids must be list or tuple or set, but received %s'
-            % type(type(name_ids))
-        )
-
-    def create_node_for_name(name):
-        if '.' not in name:
-            return gast.Name(
-                id=name, ctx=ctx, annotation=None, type_comment=None
-            )
-        return gast.parse(name).body[0].value
-
-    gast_names = [create_node_for_name(name_id) for name_id in name_ids]
-    if len(gast_names) == 1 and not gen_tuple_if_single:
-        name_node = gast_names[0]
-    else:
-        name_node = gast.Tuple(elts=gast_names, ctx=ctx)
-    return name_node
-
-
-def create_funcDef_node(nodes, name, input_args, return_name_ids):
-    """
-    Wrapper all statements of nodes into one ast.FunctionDef, which can be
-    called by ast.Call.
-    """
-    nodes = copy.copy(nodes)
-    # add return statement
-    if return_name_ids:
-        nodes.append(gast.Return(value=generate_name_node(return_name_ids)))
-    else:
-        nodes.append(gast.Return(value=None))
-    func_def_node = gast.FunctionDef(
-        name=name,
-        args=input_args,
-        body=nodes,
-        decorator_list=[],
-        returns=None,
-        type_comment=None,
-    )
-    return func_def_node
-
-
-def create_assign_node(name, node):
-    """
-    Creates a `gast.Assign` node by given name_id as target and node as value.
-    """
-    targets = generate_name_node(name, ctx=gast.Store())
-    assign_node = gast.Assign(targets=[targets], value=node)
-    return targets, assign_node
 
 
 def get_temp_dir():
@@ -1084,17 +950,6 @@ class GetterSetterHelper:
         for i, v in zip(indices, values):
             vars[i] = v
         self.setter(vars)
-
-
-def create_name_str(name_ids):
-    """
-    Return "('x', 'y')" for [x, y]
-    """
-    if not name_ids:
-        return 'None'
-
-    names_str = ["'%s'" % (name.replace("'", "\\'")) for name in name_ids]
-    return "(%s, )" % ','.join(names_str)
 
 
 def prim_or_cinn_is_enabled(build_strategy, backend):
